Add heading check
This is typically not something that is in a manual, but rather taught
in the certification class an officer takes to use radar (in the USA).
The further off the road they sit to run radar, or the more of an angle
that sit to traffic, the less accurate the return will be. Instead of
requiring users to know how doppler works, simplify it and just not
return the vehicle.

The original idea stemmed from https://www.quora.com/How-does-police-radar-work-when-officers-are-perpendicular-to-the-road