fix: audio not playing on speed lock
This was caused because I used the RADAR:IsMenuOpen() function to control if audio should be played. This is so if one player is in the operator menu, and the other locks a speed, the audio is not played for the player in the operator menu. The variable 'isMenuOpen' just needed inverting.