fix: key binds not registering until resource restart
There were a number of people with issues regarding their key binds not getting registered until the resource was restarted. This appears to have been caused by QBus interfering with the `playerSpawned` event which the resource relied on to register the key binds. The exact qb resource in question is unknown presently.